[SYNC-48][FE] 🚚 move CitationManager to components/Editor
