+++ src/components/Editor/CitationManager/index.js
